initial

+++ cifar100_224x224_ppio.py
@@ -91,7 +91,8 @@ class ResNet50Classifier(pl.LightningModule):
 def main():
     data_module = CIFAR100DataModule(batch_size=256)
     model = ResNet50Classifier()
-    trainer = Trainer(accelerator='gpu', devices=1, max_epochs=5, precision=16)
+    # trainer = Trainer(accelerator='gpu', devices=1, max_epochs=5, precision=16)
+    trainer = Trainer(accelerator='gpu', devices=1, max_epochs=5)
     trainer.fit(model, datamodule=data_module)
     trainer.test(model, datamodule=data_module)
